requirements.txt
requests
beautifulsoup4
lxml
streamlit
schedule
python-dotenv
pandas
matplotlib
pytest
